All tests pass, because we added more classes to handle other sections
Added a Inverted threading, and a right threading to handle those cases that were not covered by normal threading class. changed the setResult method in MinimumResult class so only check if the result is greater than the new result, if so, then result is the new result. fixed issue where private long was not the correct value on a empty list. Then comments were added.